Install script for v2.7 works and tested.

+++ TerrainLOS_Tests/test_terrainLOS.py
@@ -6,15 +6,17 @@
 #
 # Author: Sam Mansfield
 
+import sys
 import os
 import subprocess
+import shutil
 
 def print_usage():
   print("Correct usage:")
   print("  python test_terrainLOS.py path_to_contiki")
   exit()
 
-if len(os.argv) != 2:
+if len(sys.argv) != 2:
   print_usage()
 
 starting_directory = os.getcwd()
@@ -25,12 +27,23 @@ contiki_path = sys.argv[1]
 if contiki_path[-1] != "/":
   contiki_path += "/"
 
-os.chdir(contiki + "tools/cooja")
-os.remove("build/dag.xml")
-output = subprocess.check_output(["ant", "run_nogui", "-Dargs=" + starting_directory + "terrainLOS_simple_well_test.csc"])
+os.chdir(contiki_path + "tools/cooja")
+if os.path.exists("build/dag.xml"):
+  os.remove("build/dag.xml")
+
+print("Running Simple Well Test")
+output = subprocess.check_output(["ant", "run_nogui", "-Dargs=" + starting_directory + "terrainLOS_simple_well_test_2_7.csc"])
+##print(output)
+
 shutil.copyfile("build/dag.xml", starting_directory + "dag.xml")
 os.chdir(starting_directory)
-output = subprocess.check_output(["diff", "terrainLOS_simple_well_test_dag.xml", "dag.xml"])
+
+# Diff only exits on 0 if files match
+try:
+  output = subprocess.check_output(["diff", "terrainLOS_simple_well_test_dag_2_7.xml", "dag.xml"])
+except subprocess.CalledProcessError as e:
+  output = e.output
+   
 if output == "":
   print("PASSED")
 else:
